Split connection_id tests
