+++ test/integration/CMakeLists.txt
@@ -27,6 +27,7 @@ add_executable(
     test/reconnect.cpp
     test/any_connection.cpp
     test/character_set_tracking.cpp
+    test/connection_id.cpp
     test/pipeline.cpp
     test/connection_pool.cpp
     test/db_specific.cpp
